requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
anthropic>=0.34.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
websockets==12.0

